DMP-556 Phase 1 refactoring
