+++ src/integrationTest/java/uk/gov/hmcts/darts/dailylist/controller/DailyListProcessEndpointTest.java
@@ -10,7 +10,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import uk.gov.hmcts.darts.common.entity.CourthouseEntity;
 import uk.gov.hmcts.darts.common.exception.DartsApiException;
-import uk.gov.hmcts.darts.courthouse.CourthouseRepository;
+import uk.gov.hmcts.darts.common.repository.CourthouseRepository;
 import uk.gov.hmcts.darts.dailylist.service.DailyListProcessor;
 
 import java.time.LocalDate;
