+++ src/main/java/uk/gov/hmcts/darts/common/repository/CommonCourthouseRepository.java
@@ -4,7 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import uk.gov.hmcts.darts.common.entity.CourthouseEntity;
 
+import java.util.Optional;
+
 @Repository
 public interface CommonCourthouseRepository extends JpaRepository<CourthouseEntity, Integer> {
-    CourthouseEntity findByCourthouseNameIgnoreCase(String name);
+    Optional<CourthouseEntity> findByCourthouseNameIgnoreCase(String name);
 }
