+++ src/main/java/uk/gov/hmcts/darts/common/repository/CourthouseRepository.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.darts.courthouse;
+package uk.gov.hmcts.darts.common.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
