+++ src/main/java/uk/gov/hmcts/darts/common/repository/MediaRequestRepository.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.darts.audio.repository;
+package uk.gov.hmcts.darts.common.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
